I would have liked to hope that `p.done()`, after inlining, would eliminate the bounds check on `p.c()`. That does not seem to have been the case.

This inlines all `p.done()` checks that led to `p.c()` which leads to the majority of this speed boost.

Another chunk of speed boost came from adding a "fast path" `c <= ' '` check to `isSpace`. I initially thought that four conditionals could be pipelined, but it's still better to just have one conditional for the normal case.

Some other code is reorganized to be more condensed.

Compact no longer re-evaluates the beginning-quote index in `beginStrOrEmpty`.

I played around with removing all `p.done()` calls but what remains (all in number parsing) did not change the ns bench time on benchmarks.

```
name       old time/op  new time/op  delta
Valid-4     456ns ± 0%   325ns ± 0%  -28.68%  (p=0.000 n=9+9)
Compact-4   507ns ± 0%   398ns ± 0%  -21.50%  (p=0.000 n=7+10)
```